remove generated patch file from the repo

+++ patch.sh
@@ -1,25 +1,25 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 cd "$(dirname "$0")"
 source ./version
 
 echo "==> getting sources"
 
-if [ ! -d 'st' ] || [ ! -d 'st/.git' ]; then
-  rm -rfv 'st'
-  git clone git://git.suckless.org/st st
-  cd 'st'
-  git checkout $rev
+[ -e 'st' ] && rm -rfv 'st'
 
-else
-  cd 'st'
-  git reset --hard origin/master
-  git checkout master
-  git clean -xdf
-  git pull
-  git checkout $rev
-fi
+git clone $url 'st' && pushd 'st'
+git checkout $rev
 
-echo "==> applying unified patch file"
+echo "==> applying patches"
 
-git am ../st-community-*-$rev.diff
+for file in ../patches/*.diff; do
+  if ! patch -p1 --no-backup-if-mismatch < $file; then
+    echo "=> failed to apply patch $(basename $file), aborting..."
+    exit
+
+  else
+    echo "=> correctly applied patch $(basename $file)"
+  fi
+done
+
+popd
